refactor: finalize clustered agent naming renames and restore installation_scripts for SDK compatibility
