+++ agent_soc_manager/tools/chatops_tools.py
@@ -5,7 +5,7 @@ import os
 import httpx
 from google.adk.agents.context import Context
 
-from soc_agent.tools.chatops.card_client import generate_action_url
+from agent_soc_manager.tools.chatops.card_client import generate_action_url
 
 
 logger = logging.getLogger(__name__)
@@ -109,7 +109,7 @@ async def dispatch_card(template_name: str, ctx: Context, **kwargs) -> str:
     Dynamically loads and dispatches a Google Chat card from the chatops library.
 
     Args:
-        template_name: The name of the python module in soc_agent.tools.chatops (without .py)
+        template_name: The name of the python module in agent_soc_manager.tools.chatops (without .py)
         ctx: ADK Context to extract session and user IDs
         **kwargs: Dynamic variables to pass to the card's get_card() function.
     """
@@ -121,7 +121,7 @@ async def dispatch_card(template_name: str, ctx: Context, **kwargs) -> str:
         if chatops_path not in sys.path:
             sys.path.insert(0, chatops_path)
 
-        module = importlib.import_module(f"soc_agent.tools.chatops.{template_name}")
+        module = importlib.import_module(f"agent_soc_manager.tools.chatops.{template_name}")
         if not hasattr(module, "get_card"):
             return f"Error: Template '{template_name}' does not implement get_card()."
