+++ installation_scripts/manage_chat_ops.py
