+++ installation_scripts/manage_gcs.py
