+++ installation_scripts/manage_models.py
