+++ installation_scripts/manage_oauth.py
